Switch back to nyc /w @istanbuljs/esm-loader-hook
c8 relies on V8 coverage which isn't as good as nyc.
Empty lines count as covered, which improves coverage percentage
depending on the number of empty lines in a file.
Also it doesn't detect an uncovered else-path when there is only an
if-clause without else.

This requires Node.js v16.17.0, which adds support for multiple loaders
(@istanbuljs/esm-loader-hook + esmock).