[BREAKING] Replace npmlog with @ui5/logger/Logger
Remove the usage of npmlog and refactor @ui5/logger moduels to emit log
events which are then catched in dedicated handlers. This is somewhat
inspired by npm's proc-log module.

Internally, there are additional logger-classes for certain use cases.
Like updating status information on start/end  of a project or -task
build. This allows handlers to reate metrics on the build progress and,
for example, render a progress bar.

All of this should make @ui5/logger more future-proof. Thanks to the
decoupling of log-message creation and -handling, multiple intances and
versions of @ui5/logger can interoperate with each other.